Customer validation together with retrieval, eradication --> enhancement

+++ BACK/Delivery/BACK/Controller/Customer.js
@@ -3,47 +3,50 @@ const customer = require('../Model/Customer.js').customer;
 const JWT = require('jsonwebtoken');
 const mystery = require('../Confidential/JWT.js').mystery;  // Confidential key
 
-const extraction = async ( request , response ) => {  // Faith --> Datum extraction provided, customer unique identifier
+const retrieval = async ( request , response ) => {  // Faith --> Absolute database datum, retrieval
     try{
-        const identifier = request.params.identifier;
-        const database_response = await customer.findById( identifier );  // Return document provided, particular unique identifier
-
-        if( database_response != null ){  // Authentic, customer identifier
-            response.json( { metadata : database_response } )
+        if( request.responsibility == 'Administrator' ){
+            const database_response = await customer.find();  // Faith --> Absolute database datum, retrieval
+            if( database_response != null ){  // Authentic, customer identifier
+                response.json( { essence : 'Successful, datum retrieval' , metadata : database_response } )
+            }else{
+                response.status( 502 );  // Bad gate-way
+                response.json( { essence : 'Blunder, bad gate-way' } );
+            }
         }else{
             response.status( 401 );  // Represent, corrupt credential
-            response.json( { essence : 'Corrupt credential' } );
+            response.json( { essence : 'Access prohibited' } );
         }
+    }catch( blunder ){
+        response.status( 500 );  // Server blunder
+        response.json( { essence : blunder.message } )
+    }
+}
+
+const extraction = async ( request , response ) => {  // Faith --> Datum extraction provided, customer unique identifier
+    try{
+        response.json( { metadata : request.metadata } )
     }
     catch( blunder ){
-        response.status( 502 );  // Bad gate-way
-        response.json( { essence : blunder.message } );
+        response.status( 500 );  // Server blunder
+        response.json( { essence : blunder.message } )
     }
 }
 
 const mutation = async ( request , response ) => {  // Faith --> Mutate datum
     try{
-        const identifier = request.params.identifier;  // Represent, customer identifier
-        const database_response = await customer.findById( identifier );  // Return document provided, particular unique identifier
+        const datum = request.body;
+        for ( const key in datum ) {  // Looping through, specified datum
+            request.metadata[ key ] = datum[ key ];  // Mutation
+        }
 
-        if( database_response != null ){  // Authentic, customer identifier
-            const datum = request.body;
-            for ( const key in datum ) {  // Looping through, specified datum
-                database_response[ key ] = datum[ key ];  // Mutation
-            }
+        const translation = await request.metadata.save();  // 'Save', translated document
 
-            const translation = await database_response.save();  // 'Save', translated document
-
-            if( translation != undefined ){
-                response.json( { essence : 'Successful, mutation' , metadata : translation } )
-            }else{
-                response.status( 502 );  // Bad gate-way
-                response.json( { essence : 'Blunder, bad gate-way' } );
-            }
-
+        if( translation ){
+            response.json( { essence : 'Successful, mutation' , metadata : translation } )
         }else{
-            response.status( 401 );  // Represent, corrupt credential
-            response.json( { essence : 'Corrupt credential' } );
+            response.status( 502 );  // Bad gate-way
+            response.json( { essence : 'Blunder, bad gate-way' } );
         }
     }
     catch( blunder ){
@@ -54,14 +57,24 @@ const mutation = async ( request , response ) => {  // Faith --> Mutate datum
 
 const eradication = async ( request , response ) => {  // Faith --> Resource eradication
     try{
-        const identifier = request.params.identifier;
-        const database_response = await customer.findByIdAndDelete( identifier );  // Document eradication
+        if( request.params.identifier ){
+            const identifier = request.params.identifier;
+            if( identifier == request.identifier ){
+                const database_response = await customer.findByIdAndDelete( identifier );  // Document eradication
 
-        if( database_response != null ){  // Authentic, customer identifier
-            response.json( { essence : 'Successful, document eradication' , metadata : database_response } )
+                if( database_response != null ){  // Authentic, customer identifier
+                    response.json( { essence : 'Successful, document eradication' , metadata : database_response } )
+                }else{
+                    response.status( 401 );  // Represent, corrupt credential
+                    response.json( { essence : 'Corrupt credential' } );
+                }
+            }else{
+                response.status( 401 );  // Represent, corrupt credential
+                response.json( { essence : 'Authenticate' } );
+            }
         }else{
             response.status( 401 );  // Represent, corrupt credential
-            response.json( { essence : 'Corrupt credential' } );
+            response.json( { essence : 'Authenticate' } );
         }
     }
     catch( blunder ){
@@ -70,7 +83,7 @@ const eradication = async ( request , response ) => {  // Faith --> Resource era
     }
 }
 
-const validation = async ( request , response , future ) => {
+const validation = async ( request , response , future ) => {  // Validation, provided cookie
     try {
         const cookie = request.cookies;  // Represent cookie, holding up JWT
     
@@ -78,15 +91,13 @@ const validation = async ( request , response , future ) => {
             const payload = JWT.verify( cookie.Authentication , mystery );  // Represent, customer identifier
             
             if( payload ){
-                if( request.params.identifier == payload ){  // Access modification
-                    const database_metadata = await customer.findById( payload.payload );
+                const database_metadata = await customer.findById( payload.payload );
 
-                    if( database_metadata ){  // Represent, authentic customer
-                        future();
-                    }else{
-                        response.status( 401 );  // Represent, corrupt credential
-                        response.json( { essence : 'Authenticate' } );
-                    }
+                if( database_metadata ){  // Represent, authentic customer
+                    request.identifier = payload.payload
+                    request.metadata = database_metadata;  // Maintainance
+                    request.responsibility = database_metadata.responsibility; 
+                    future();
                 }else{
                     response.status( 401 );  // Represent, corrupt credential
                     response.json( { essence : 'Authenticate' } );
@@ -105,4 +116,4 @@ const validation = async ( request , response , future ) => {
     }
 }
 
-module.exports = { extraction , mutation , eradication , validation };
+module.exports = { retrieval , extraction , mutation , eradication , validation };
